+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 .DS_Store
 z ex.
 bundle.js
+datafile.json
