+++ db/seedData/seed.1.js
@@ -1,5 +1,5 @@
 const faker = require('faker');
-const Product = require('./index.js');
+const Product = require('../index.js');
 const fs = require('fs');
 const stream = require('stream'); 
 
@@ -9,7 +9,7 @@ const saveProducts = function() {
   let fd = fs.openSync('datafile.json', 'a');
   let data = [];
   let datacount = 0;
-  for (let i = 1; i <= 100000; i++) {
+  for (let i = 1; i <= 1000000; i++) {
     console.time('datagroup')
     let instance = new Product({
       projectID: i,
